Moved `ContextPeopleExt`, `ContextPeopleExtInternal`, and related tests to submodule `people::context_extension`.
